Added global singleton, shell for winCondition  & playerCaught
The globals is a script attached to a same named object and tagged the same. winCondition is to be called to check if the player has won, and to end the game if they have. playerCaught is for the demon to call and will handle ending the game. Both functions are shells ATM